refactor(daemon): keep migration scratch writes in the projection boundary
